Move the abstract builder class to the package level
